DKAFC-1397:  Supply the RLMM security configs to tiered_replication system test.
Summary:
This replication system test runs for both ZK and KRaft modes. ZK mode has 31 out of 49 tests and remaining 18 are for KRaft mode.  All the 31 tests are passed now.

Disabled the KRaft mode tests, as there are a lot of changes happened between v2.8.0 and v3.1.0 and it's in beta mode. We can turn on the KRaft mode when merging this patch to the trunk.

Test Plan:
PFA system test report:

{F613203629}

{F613203627}

Reviewers: satishd, prateeka

Reviewed By: satishd

JIRA Issues: DKAFC-1397

Differential Revision: https://code.uberinternal.com/D7398069