DKAFC-2308: Verify that the follower fetch requests are handled when fetch-offsets are out-of-range
Summary:
Added integration test to verify that the follower fetch requests are handled correctly when the given fetch-offsets are out-of-range.

This case can happen when all the remote log segments are deleted for a partition. Then, when starting a failed broker with the new disk, it should not query the remote log storage and only replicate the data that is there in the leader.

Test Plan: IT

Reviewers: satishd, prateeka

Reviewed By: satishd

JIRA Issues: DKAFC-2308

Differential Revision: https://code.uberinternal.com/D7662555